backend/requirements.txt
requests
flake8
black
Flask
fastapi
uvicorn
Pillow

python-multipart
